+++ crates/mocks/src/backend.rs
@@ -1,6 +1,6 @@
 use std::{cell::UnsafeCell, collections::BTreeMap};
 
-use storey_storage::IterableStorage as _;
+use storey_storage::{IterableStorage, RevIterableStorage, StorageBackend, StorageBackendMut};
 
 // `UnsafeCell` is needed here to implement interior mutability.
 // https://doc.rust-lang.org/book/ch15-05-interior-mutability.html
@@ -31,14 +31,14 @@ impl Default for TestStorage {
 // Moreover, we can further guarantee that the dereference is valid because the data
 // is always initialized during construction.
 
-impl storey_storage::StorageBackend for TestStorage {
+impl StorageBackend for TestStorage {
     fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
         // Safety: see above
         unsafe { (*self.0.get()).get(key).cloned() }
     }
 }
 
-impl storey_storage::StorageBackendMut for TestStorage {
+impl StorageBackendMut for TestStorage {
     fn set(&mut self, key: &[u8], value: &[u8]) {
         // Safety: see above
         unsafe {
@@ -54,7 +54,7 @@ impl storey_storage::StorageBackendMut for TestStorage {
     }
 }
 
-impl storey_storage::IterableStorage for TestStorage {
+impl IterableStorage for TestStorage {
     type KeysIterator<'a> = Box<dyn DoubleEndedIterator<Item = Vec<u8>> + 'a>;
     type ValuesIterator<'a> = Box<dyn DoubleEndedIterator<Item = Vec<u8>> + 'a>;
     type PairsIterator<'a> = Box<dyn DoubleEndedIterator<Item = (Vec<u8>, Vec<u8>)> + 'a>;
@@ -98,7 +98,7 @@ impl storey_storage::IterableStorage for TestStorage {
     }
 }
 
-impl storey_storage::RevIterableStorage for TestStorage {
+impl RevIterableStorage for TestStorage {
     type RevKeysIterator<'a> = Box<dyn Iterator<Item = Vec<u8>> + 'a>;
     type RevValuesIterator<'a> = Box<dyn Iterator<Item = Vec<u8>> + 'a>;
     type RevPairsIterator<'a> = Box<dyn Iterator<Item = (Vec<u8>, Vec<u8>)> + 'a>;
@@ -140,4 +140,122 @@ fn check_bounds(v: &[u8], start: Option<&Vec<u8>>, end: Option<&Vec<u8>>) -> boo
         }
     }
     true
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn storage_backend() {
+        // TODO: split this into smaller tests?
+
+        let mut storage = TestStorage::new();
+
+        storage.set(&[0], b"bar");
+        storage.set(&[1], b"baz");
+        storage.set(&[1, 0], b"qux");
+        storage.set(&[1, 1], b"quux");
+        storage.set(&[2], b"qux");
+
+        let keys: Vec<_> = storage.keys(None, None).collect();
+        assert_eq!(
+            keys,
+            vec![vec![0], vec![1], vec![1, 0], vec![1, 1], vec![2]]
+        );
+
+        let some_keys: Vec<_> = storage.keys(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(some_keys, vec![vec![1], vec![1, 0], vec![1, 1]]);
+
+        let values: Vec<_> = storage.values(None, None).collect();
+        assert_eq!(
+            values.iter().collect::<Vec<_>>(),
+            vec![&b"bar"[..], b"baz", b"qux", b"quux", b"qux"]
+        );
+
+        let some_values: Vec<_> = storage.values(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(
+            some_values.iter().collect::<Vec<_>>(),
+            vec![&b"baz"[..], b"qux", b"quux"]
+        );
+
+        let pairs: Vec<_> = storage.pairs(None, None).collect();
+        assert_eq!(
+            pairs,
+            vec![
+                (vec![0], b"bar".to_vec()),
+                (vec![1], b"baz".to_vec()),
+                (vec![1, 0], b"qux".to_vec()),
+                (vec![1, 1], b"quux".to_vec()),
+                (vec![2], b"qux".to_vec()),
+            ]
+        );
+
+        let some_pairs: Vec<_> = storage.pairs(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(
+            some_pairs,
+            vec![
+                (vec![1], b"baz".to_vec()),
+                (vec![1, 0], b"qux".to_vec()),
+                (vec![1, 1], b"quux".to_vec()),
+            ]
+        );
+
+        let rev_keys: Vec<_> = storage.rev_keys(None, None).collect();
+        assert_eq!(
+            rev_keys,
+            vec![vec![2], vec![1, 1], vec![1, 0], vec![1], vec![0]]
+        );
+
+        let some_rev_keys: Vec<_> = storage.rev_keys(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(some_rev_keys, vec![vec![1, 1], vec![1, 0], vec![1]]);
+
+        let rev_values: Vec<_> = storage.rev_values(None, None).collect();
+        assert_eq!(
+            rev_values.iter().collect::<Vec<_>>(),
+            vec![&b"qux"[..], b"quux", b"qux", b"baz", b"bar"]
+        );
+
+        let some_rev_values: Vec<_> = storage.rev_values(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(
+            some_rev_values.iter().collect::<Vec<_>>(),
+            vec![&b"quux"[..], b"qux", b"baz"]
+        );
+
+        let rev_pairs: Vec<_> = storage.rev_pairs(None, None).collect();
+        assert_eq!(
+            rev_pairs,
+            vec![
+                (vec![2], b"qux".to_vec()),
+                (vec![1, 1], b"quux".to_vec()),
+                (vec![1, 0], b"qux".to_vec()),
+                (vec![1], b"baz".to_vec()),
+                (vec![0], b"bar".to_vec()),
+            ]
+        );
+
+        let some_rev_pairs: Vec<_> = storage.rev_pairs(Some(&[1]), Some(&[2])).collect();
+        assert_eq!(
+            some_rev_pairs,
+            vec![
+                (vec![1, 1], b"quux".to_vec()),
+                (vec![1, 0], b"qux".to_vec()),
+                (vec![1], b"baz".to_vec()),
+            ]
+        );
+    }
+
+    #[test]
+    fn metadata() {
+        use storey_storage::StorageMut as _;
+
+        let mut storage = TestStorage::new();
+        storage.set_meta(&[0], b"meta");
+
+        assert_eq!(StorageBackend::get(&storage, &[0]), None);
+        assert_eq!(
+            StorageBackend::get(&storage, &[255, 0]),
+            Some(b"meta".to_vec())
+        );
+    }
 }
