+++ crates/mocks/src/encoding.rs
@@ -54,3 +54,18 @@ impl MyTestEncoding for u64 {
         Ok(u64::from_le_bytes(bytes))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use storey_encoding::{DecodableWith as _, EncodableWith as _};
+
+    #[test]
+    fn encoding() {
+        assert_eq!(12u64.encode(), Ok(12u64.to_le_bytes().to_vec()));
+    }
+
+    #[test]
+    fn decoding() {
+        assert_eq!(<u64>::decode(&12u64.to_le_bytes()), Ok(12));
+    }
+}
